1. add Error 2. add checking file existing 3. refactoring

+++ app/src/main/java/ru/net/serbis/share/task/DeleteTask.java
@@ -4,8 +4,9 @@ import android.os.*;
 import jcifs.smb.*;
 import ru.net.serbis.share.*;
 import ru.net.serbis.share.tool.*;
+import ru.net.serbis.share.data.Error;
 
-public class DeleteTask extends AsyncTask<String, Void, String>
+public class DeleteTask extends AsyncTask<String, Void, Error>
 {
     private BrowserCallback callback;
     private Smb smb;
@@ -17,24 +18,30 @@ public class DeleteTask extends AsyncTask<String, Void, String>
     }
 
     @Override
-    protected String doInBackground(String... params)
+    protected Error doInBackground(String... params)
     {
         try
         {
             String path = params[0];
             SmbFile source = smb.getFile(path);
+
+            if (!source.isFile() || !source.exists())
+            {
+                return new Error(smb, Constants.ERROR_FILE_IS_NOT_FOUND, R.string.error_file_is_not_found);
+            }
+
             source.delete();
             return null;
         }
         catch (Throwable e)
         {
             Log.error(this, e);
-            return e.getMessage();
+            return new Error(Constants.ERROR_DELETE, e.getMessage());
         }
     }
 
     @Override
-    protected void onPostExecute(String error)
+    protected void onPostExecute(Error error)
     {
         if (error == null)
         {
@@ -42,7 +49,7 @@ public class DeleteTask extends AsyncTask<String, Void, String>
         }
         else
         {
-            callback.onError(Constants.ERROR_DELETE, error);
+            callback.onError(error);
         }
     }
 }
